pyproject.toml
[project]
name = "townspark-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-jazzmin>=3.0.1",
    "django-stubs>=5.2.8",
    "django-unfold>=0.74.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.6",
    "djoser>=2.3.3",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "uuid>=1.30",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main_app.settings"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
addopts = "-s"